requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
Pillow>=9.5.0,<9.6
django-summernote>=0.8.20.0,<0.8.21
python-dotenv>=1.0.0,<1.1
django-axes>=6.0.1,<6.1
django-crispy-forms
crispy-bootstrap4
django-tables2
django-widget-tweaks
requests
django-filter